Remove Bootstrap
I want to support proper dark mode. And I want to do so with CSS variables. And Bootstrap and CSS variables are in a pretty weird state. So… just … do it all manually. This also lets me get a little more finicky with some of the behaviors (like a search bar in the faction dropdown with extremely specific keyboard handling to be _juuuust_ how I want it).

Things of note:
* Faction colors are largely migrated to CSS variables and away from fully-inline styles
* Faction colors now have an auto-computed hover color for nicer hover and active states
* As mentioned above, faction dropdown gained a search field and better keyboard handling
* Characters screen live overlays can now be interacted with
* Suggestion form has better errors. They don't show up as quickly, they clear better, there are error messages, you can dismiss by clicking away if you haven't filled anything out.
* Mobile lost the nav bar hamburger menu collapse thing. I should maybe add that back