Update PreloadedData so that we don't later fall back to stale data
If you, say, load a screen that auto reloads and eventually get new data, then navigate away and back to that screen, we would fall back to the original preloaded data. This fixes that by always replacing preloaded data when we get subsequent loads.

This also means that we effectively share loads between various pages. Which TBH I don't hate. Can't decide if it's a little _too_ hidden of a behavior though (i.e. it's share in a mutable context object rather than being part of the “state tree”). Might be nicer to have this lifted out into a proper system rather that buried in the requests themselves.