Fix useNow
Previously, preloadedContext.now was a date in quotes. Browsers seem to be more forgiving and parse a date from that, but node gave an invalid date. This meant that we were sending “NaNh NaNm” for stream runtimes that were then replaces on the next render loop, but gave a flash of incorrect data.