Fix sameYear comparison for time zone differences
Same as before, but for `relativeDateDiff`:

Two dates may be in different years in one time zone and the same year in another. This can cause
mismatches between server and client rendering. Fix it by using a localized representation of the
year for the comparison, which already handles SSR correctly.