Nicer telegram obscuring on streamer pages and mobile
Since streamer pages don’t have row hovering, it’s odd to treat the entire row as a hover target for
showing the button. So instead only revel the button when hovering the obscured telegram.

On mobile and other devices where hovering isn’t a thing, always show the button.