Less flipping of useNow values
When the `useNow` result is used as a request parameter we were always issuing an extra request: one
with the preloaded `now` and one with the immediately-updated new date. Instead, only use the
preloaded date for the initial render and use the current date as the starting value past that.